backend/requirements.txt
# PTaaS Backend Requirements

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Async Task Queue
celery==5.3.6
redis==5.0.1

# AWS/Storage
boto3==1.34.34
botocore==1.34.34

# Docker Integration
docker==7.0.0

# HTTP Client
requests==2.31.0
httpx==0.26.0

# Environment Management
python-dotenv==1.0.1

# Data Validation
email-validator==2.1.0

# Testing (optional)
pytest==7.4.4
pytest-asyncio==0.23.3

# Utils
python-multipart==0.0.6
